// backend/src/schemas/auth.schema.ts
import { z } from "zod";
import { ROLES, TOKENS } from "../constants/data";

export const registerSchema = z
  .object({
    name: z.string().min(2, "Name must be at least 2 characters long"),
    email: z.string().email("Invalid email address"),
    password: z.string().min(6, "Password must be at least 6 characters long"),
    confirmPassword: z
      .string()
      .min(6, "Confirm password must be at least 6 characters long"),
    role: z.enum(ROLES).optional(),
  })
  .refine((data) => data.password === data.confirmPassword, {
    path: ["confirmPassword"],
    message: "Passwords do not match",
  });

export const loginSchema = z.object({
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters long"),
});

export const tokenSchema = z.object({
  token: z.string().uuid(),
  type: z.enum(TOKENS),
  expiresAt: z.date(),
});